+++ loose/lib/src/convert_from_firestore.dart
@@ -1,18 +1,20 @@
 import 'package:loose/annotations.dart';
 import 'loose_exception.dart';
 
-typedef MapDecoder<T> = T Function(Map<String, Object>? value);
+typedef MapDecoder<T> = T Function(Map<String, Object> value);
 
 class FromFs {
   FromFs._();
 
-  static String string(Map<String, Object> value, {String defaultValue = ''}) {
-    return value['stringValue'] as String? ?? defaultValue;
+  static String string(Object? value, {String defaultValue = ''}) {
+    final o = value as Map<String, Object>;
+    return o['stringValue'] as String? ?? defaultValue;
   }
 
-  static String? stringNull(Map<String, Object> value,
+  static String? stringNull(Object? value,
       {String name = '', bool allowNull = false}) {
-    final v = value['stringValue'] as String?;
+    final o = value as Map<String, Object>;
+    final v = o['stringValue'] as String?;
     if (v != null) {
       return v;
     } else if (allowNull) {
@@ -23,13 +25,15 @@ class FromFs {
     }
   }
 
-  static int integer(Map<String, Object> value, {int defaultValue = 0}) {
-    return int.tryParse(value['integerValue'] as String? ?? '') ?? defaultValue;
+  static int integer(Object? value, {int defaultValue = 0}) {
+    final o = value as Map<String, Object>;
+    return int.tryParse(o['integerValue'] as String? ?? '') ?? defaultValue;
   }
 
-  static int? integerNull(Map<String, Object> value,
+  static int? integerNull(Object? value,
       {String name = '', bool allowNull = false}) {
-    final v = int.tryParse(value['integerValue'] as String? ?? '');
+    final o = value as Map<String, Object>;
+    final v = int.tryParse(o['integerValue'] as String? ?? '');
     if (v != null) {
       return v;
     } else if (allowNull) {
@@ -40,13 +44,15 @@ class FromFs {
     }
   }
 
-  static double float(Map<String, Object> value, {double defaultValue = 0.0}) {
-    return value['doubleValue'] as double? ?? defaultValue;
+  static double float(Object? value, {double defaultValue = 0.0}) {
+    final o = value as Map<String, Object>;
+    return o['doubleValue'] as double? ?? defaultValue;
   }
 
-  static double? floatNull(Map<String, Object> value,
+  static double? floatNull(Object? value,
       {String name = '', bool allowNull = false}) {
-    final v = (value['doubleValue'] as num?)?.toDouble();
+    final o = value as Map<String, Object>;
+    final v = (o['doubleValue'] as num?)?.toDouble();
     if (v != null) {
       return v;
     } else if (allowNull) {
@@ -57,13 +63,15 @@ class FromFs {
     }
   }
 
-  static bool boolean(Map<String, Object> value, {bool defaultValue = false}) {
-    return value['booleanValue'] as bool? ?? defaultValue;
+  static bool boolean(Object? value, {bool defaultValue = false}) {
+    final o = value as Map<String, Object>;
+    return o['booleanValue'] as bool? ?? defaultValue;
   }
 
-  static bool? booleanNull(Map<String, Object> value,
+  static bool? booleanNull(Object? value,
       {String name = '', bool allowNull = false}) {
-    final v = value['booleanValue'] as bool?;
+    final o = value as Map<String, Object>;
+    final v = o['booleanValue'] as bool?;
     if (v != null) {
       return v;
     } else if (allowNull) {
@@ -74,15 +82,17 @@ class FromFs {
     }
   }
 
-  static DateTime datetime(Map<String, Object> value,
+  static DateTime datetime(Object? value,
       {String defaultValue = '0000-01-01T00:00:00.000Z'}) {
-    return DateTime.tryParse(value['timestampValue'] as String? ?? '') ??
+    final o = value as Map<String, Object>;
+    return DateTime.tryParse(o['timestampValue'] as String? ?? '') ??
         DateTime.parse(defaultValue);
   }
 
-  static DateTime? datetimeNull(Map<String, Object> value,
+  static DateTime? datetimeNull(Object? value,
       {String name = '', bool allowNull = false}) {
-    final v = DateTime.tryParse(value['timestampValue'] as String? ?? '');
+    final o = value as Map<String, Object>;
+    final v = DateTime.tryParse(o['timestampValue'] as String? ?? '');
     if (v != null) {
       return v;
     } else if (allowNull) {
@@ -93,14 +103,15 @@ class FromFs {
     }
   }
 
-  static Reference reference(Map<String, Object> value,
-      {String defaultValue = '/'}) {
-    return Reference.fromFirestore(value, defaultValue);
+  static Reference reference(Object? value, {String defaultValue = ''}) {
+    final o = value as Map<String, Object>;
+    return Reference.fromFirestore(o, defaultValue);
   }
 
-  static Reference? referenceNull(Map<String, Object> value,
+  static Reference? referenceNull(Object? value,
       {String name = '', bool allowNull = false}) {
-    final v = value['referenceValue'];
+    final o = value as Map<String, Object>;
+    final v = o['referenceValue'];
     if (v != null) {
       return Reference.fromFirestore(value);
     } else if (allowNull) {
@@ -111,45 +122,46 @@ class FromFs {
     }
   }
 
-  static List<T> list<T>(Map<String, Object> value, MapDecoder<T> mapDecoder,
+  static List<T> list<T>(Object? value, MapDecoder<T> mapDecoder,
       {List<T> defaultValue = const []}) {
-    final list =
-        (value['arrayValue'] as Map<String, Object>)['values'] as List? ??
-            defaultValue;
-    return list.map((e) => mapDecoder(e as Map<String, Object>?)).toList();
+    final o = value as Map<String, Object>;
+    final list = (o['arrayValue'] as Map<String, Object>)['values'] as List? ??
+        defaultValue;
+    return list.map((e) => mapDecoder(e as Map<String, Object>)).toList();
   }
 
-  static List<T>? listNull<T>(
-      Map<String, Object> value, MapDecoder<T> mapDecoder,
+  static List<T>? listNull<T>(Object? value, MapDecoder<T> mapDecoder,
       {String name = '', bool allowNull = false}) {
-    if (value == null && allowNull) {
-      return null;
-    } else if (value == null && !allowNull) {
-      final knownField = 'in field "$name" ';
-      throw LooseException(
-          'A null value was read ${knownField}but nulls are not allowed. Use either "allowNull" or "readNull" annotation to allow.');
-    }
-    if (value.containsKey('arrayValue')) {
+    final o = value as Map<String, Object>;
+    if (o.containsKey('arrayValue')) {
       final valuesList =
-          (value['arrayValue'] as Map<String, Object>)['values'] as List? ??
+          (o['arrayValue'] as Map<String, Object>)['values'] as List? ??
               const [];
       return valuesList
           .map((e) => mapDecoder(e as Map<String, Object>))
           .toList();
+    } else if (allowNull) {
+      return null;
+    } else {
+      final knownField = 'in field "$name" ';
+      throw LooseException(
+          'A null value was read ${knownField}but nulls are not allowed. Use either "allowNull" or "readNull" annotation to allow.');
     }
-    return null;
   }
 
-  static T map<T>(Map<String, Object> value, MapDecoder<T> mapDecoder,
+  static T map<T>(Object? value, MapDecoder<T> mapDecoder,
       {Map<String, Object> defaultValue = const {}}) {
-    final valueMap = (value['mapValue'] as Map<String, Object>)['fields']
-        as Map<String, Object>?;
+    final o = value as Map<String, Object>;
+    final valueMap = (o['mapValue'] as Map<String, Object>)['fields']
+            as Map<String, Object>? ??
+        const {};
     return mapDecoder(valueMap);
   }
 
-  static T? mapNull<T>(Map<String, Object> value, MapDecoder<T> mapDecoder,
+  static T? mapNull<T>(Object? value, MapDecoder<T> mapDecoder,
       {String name = '', bool allowNull = false}) {
-    final v = (value['mapValue'] as Map<String, Object>)['fields']
+    final o = value as Map<String, Object>;
+    final v = (o['mapValue'] as Map<String, Object>)['fields']
         as Map<String, Object>?;
     if (v != null) {
       return mapDecoder(v);