+++ loose/lib/src/convert_to_firestore.dart
@@ -6,14 +6,10 @@ class ToFs {
 
   static const _toNullValue = {'nullValue': 'NULL_VALUE'};
 
-  static Map<String, Object> string(String string, String name,
+  static Map<String, Object> string(String? string, String name,
       {String defaultValue = '',
       bool useDefaultValue = false,
       bool allowNull = false}) {
-    if (useDefaultValue && allowNull) {
-      throw LooseException(
-          'Cannot allow null and use default value for "$name". Must only use one or neither.');
-    }
     if (string != null) {
       return {'stringValue': string};
     } else if (allowNull) {
@@ -25,14 +21,10 @@ class ToFs {
     }
   }
 
-  static Map<String, Object> integer(int integer, String name,
+  static Map<String, Object> integer(int? integer, String name,
       {int defaultValue = 0,
       bool useDefaultValue = false,
       bool allowNull = false}) {
-    if (useDefaultValue && allowNull) {
-      throw LooseException(
-          'Cannot allow null and use default value for "$name". Must only use one or neither.');
-    }
     if (integer != null) {
       return {'integerValue': integer.toString()};
     } else if (allowNull) {
@@ -44,14 +36,10 @@ class ToFs {
     }
   }
 
-  static Map<String, Object> float(double float, String name,
+  static Map<String, Object> float(double? float, String name,
       {double defaultValue = 0.0,
       bool useDefaultValue = false,
       bool allowNull = false}) {
-    if (useDefaultValue && allowNull) {
-      throw LooseException(
-          'Cannot allow null and use default value for "$name". Must only use one or neither.');
-    }
     if (float != null) {
       return {'doubleValue': float};
     } else if (allowNull) {
@@ -63,14 +51,10 @@ class ToFs {
     }
   }
 
-  static Map<String, Object> boolean(bool boolean, String name,
+  static Map<String, Object> boolean(bool? boolean, String name,
       {bool defaultValue = false,
       bool useDefaultValue = false,
       bool allowNull = false}) {
-    if (useDefaultValue && allowNull) {
-      throw LooseException(
-          'Cannot allow null and use default value for "$name". Must only use one or neither.');
-    }
     if (boolean != null) {
       return {'booleanValue': boolean};
     } else if (allowNull) {
@@ -82,14 +66,10 @@ class ToFs {
     }
   }
 
-  static Map<String, Object> datetime(DateTime datetime, String name,
+  static Map<String, Object> datetime(DateTime? datetime, String name,
       {String defaultValue = '0000-01-01T00:00:00.000Z',
       bool useDefaultValue = false,
       bool allowNull = false}) {
-    if (useDefaultValue && allowNull) {
-      throw LooseException(
-          'Cannot allow null and use default value for "$name". Must only use one or neither.');
-    }
     if (datetime != null) {
       return {'timestampValue': datetime.toUtc().toIso8601String()};
     } else if (allowNull) {
@@ -102,14 +82,10 @@ class ToFs {
     }
   }
 
-  static Map<String, Object> reference(Reference reference, String name,
+  static Map<String, Object> reference(Reference? reference, String name,
       {String defaultValue = '/',
       bool useDefaultValue = false,
       bool allowNull = false}) {
-    if (useDefaultValue && allowNull) {
-      throw LooseException(
-          'Cannot allow null and use default value for "$name". Must only use one or neither.');
-    }
     if (reference != null && reference.name.isNotEmpty) {
       return {'referenceValue': reference.name};
     } else if (allowNull) {
@@ -121,14 +97,11 @@ class ToFs {
     }
   }
 
-  static Map<String, Object> list(List<Map<String, Object>> values, String name,
+  static Map<String, Object> list(
+      List<Map<String, Object>>? values, String name,
       {List<Map<String, Object>> defaultValue = const [],
       bool useDefaultValue = false,
       bool allowNull = false}) {
-    if (useDefaultValue && allowNull) {
-      throw LooseException(
-          'Cannot allow null and use default value for "$name". Must only use one or neither.');
-    }
     if (values != null) {
       return {
         'arrayValue': {'values': values}
@@ -144,7 +117,7 @@ class ToFs {
     }
   }
 
-  static Map<String, Object> map(Map<String, Object> fields, String name,
+  static Map<String, Object> map(Map<String, Object>? fields, String name,
       {Map<String, Object> defaultValue = const {},
       bool useDefaultValue = false,
       bool allowNull = false}) {