+++ loose_builder/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Loose Builder Changelog
 
+## 0.1.0-nullsafety.1
+- Cleanup from null safety migration
+- Fix empty Reference constant
+
 ## 0.1.0-nullsafety.0
 - Upgrade to null safety
 - Several API breaking changes